local code refactoring
