+++ src/util/subscription-set/subscription-set.ts
@@ -21,13 +21,13 @@ export class SubscriptionSetFactory {
   private cacheType: AdapterConfig['CACHE_TYPE']
   private redisClient?: Redis
   private adapterName?: string
-  private capacity: AdapterConfig['CACHE_MAX_SUBSCRIPTIONS']
+  private capacity: AdapterConfig['SUBSCRIPTION_SET_MAX_ITEMS']
 
   constructor(config: AdapterConfig, adapterName: string, redisClient?: Redis) {
     this.cacheType = config.CACHE_TYPE
     this.redisClient = redisClient
     this.adapterName = adapterName
-    this.capacity = config.CACHE_MAX_SUBSCRIPTIONS
+    this.capacity = config.SUBSCRIPTION_SET_MAX_ITEMS
   }
 
   buildSet<T>(endpointName: string): SubscriptionSet<T> {
@@ -40,7 +40,7 @@ export class SubscriptionSetFactory {
         }
         // Identifier key used for the subscription set in redis
         const subscriptionSetKey = `${this.adapterName}-${endpointName}-subscriptionSet`
-        return new RedisSubscriptionSet<T>(this.redisClient, subscriptionSetKey, this.capacity)
+        return new RedisSubscriptionSet<T>(this.redisClient, subscriptionSetKey)
       }
     }
   }
